Fix colon not displaying correctly in title
The tour's title format uses semicolon which breaks the parsing
of the generator. To avoid this we escape the colon which in turn
leads to it not being displayed correctly in title as it's not
rendered by the browser.

Replace the escaped colon for the title via regex. A better and
long term solution is to patch the generator, yet this solution is
good enough for the time being without touching any dependency.