Don’t show pointer cursor on non-clickable elements
This changes the cursor for the navigation bar. Before this change, the cursor changes to “pointer”, which normally indicates an element with a click action, as soon as you hover over a navigation “tab”, even when you’re not hovering over the link. This means that you get the “pointer” cursor in locations where clicking won’t actually do anything, which is confusing to the user, and annoying once they figure it out.

Fixed by completely removing the custom cursor. For the link, the browser will use the “pointer” cursor anyways.