Improve anchor links
Instead of linking to the full URL + anchor, just link to the anchor. No
need to reload the site for this. I suggested this in #432 and then
didn’t do anything about it for months, and I don’t even know why I
implemented the full URL version in #309 in the first place – you can
see that the actual fix, aside from a comment update, is purely to
remove some code – but anyways, it’s better now, so there we are.