Update modules chapter of tour
The introductory paragraph of the “Packages and imports” section dealt
mostly with packages, not with imports, and went into a lot of detail.
It has been updated to give a more high-level overview of both aspects,
and leave the detail to its subsections.

The introduction to packages did not make it clear that the file name
has no significance despite the convention to name files after their
content. It has been rewritten to make the relation between
declarations, source files, and packages more clear. The example has
been moved to a separate paragraph and expanded.

The section on imports now explains the import syntax a bit more.
The import examples have been expanded to include the example from the
previous section, partially merged, and comments have been added.

Fixes #473.