Clarify fictitious imported module
David Hagen pointed out on StackOverflow [1] that a module containing a
`math` package is nowhere to be found. I assume it’s intended to be
fictitious (the ceylon.math in the SDK lacks Complex), so add
org.example to the package name to clarify this and bring the import in
line with the other imports below.

Also, while we’re at it, consistently use org.example instead of
com.example.

[1]: https://stackoverflow.com/q/49592547/1420237